Add AccountRepository & AccountProfileRepository, rename Respository to Repository, change id type to ulong and refactoring

+++ MediaShop.Common/Models/Entity.cs
@@ -15,7 +15,7 @@ namespace MediaShop.Common.Models
         /// Gets or sets the identifier.
         /// </summary>
         /// <value>The identifier.</value>
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         /// <summary>
         /// Gets or sets the created date.
